fix: addressed PR review
