+++ ebl/bibliography/application/bibliography.py
@@ -6,6 +6,11 @@ from pydash import uniq_with
 from ebl.bibliography.application.duplicate_detection import (
     BibliographyDuplicateDetector,
 )
+from ebl.bibliography.application.duplicate_override import (
+    BLOCKING_DUPLICATE_DECISIONS,
+    DuplicateOverrideError,
+    validate_duplicate_override,
+)
 from ebl.bibliography.application.bibliography_repository import BibliographyRepository
 from ebl.bibliography.application.serialization import create_mongo_entry
 from ebl.bibliography.domain.reference import Reference
@@ -16,12 +21,6 @@ from ebl.users.domain.user import User
 COLLECTION = "bibliography"
 DEFAULT_EXPORT_LIMIT = 50
 MAX_EXPORT_LIMIT = 100
-BLOCKING_DUPLICATE_DECISIONS = {"likely_duplicate", "possible_duplicate"}
-MIN_DUPLICATE_OVERRIDE_REASON_LENGTH = 10
-
-
-class DuplicateOverrideError(Exception):
-    pass
 
 
 class Bibliography:
@@ -104,7 +103,7 @@ class Bibliography:
                 "No current duplicate candidates were found. "
                 "Use POST /api/v1/bibliography instead."
             )
-        self._validate_duplicate_override(override, duplicate_result)
+        validate_duplicate_override(override, duplicate_result)
         self.create(entry, user)
 
     def update_partner_entry(self, id_: str, entry: dict, user: User) -> Optional[dict]:
@@ -121,76 +120,6 @@ class Bibliography:
             if duplicate_result["decision"] in BLOCKING_DUPLICATE_DECISIONS
             else None
         )
-
-    @staticmethod
-    def _validate_duplicate_override(
-        override: Mapping[str, Any], duplicate_result: Mapping[str, Any]
-    ) -> None:
-        reason = override.get("reason")
-        if not isinstance(reason, str) or not reason.strip():
-            raise DuplicateOverrideError("override.reason must be a non-empty string.")
-
-        if len("".join(reason.split())) < MIN_DUPLICATE_OVERRIDE_REASON_LENGTH:
-            raise DuplicateOverrideError(
-                "override.reason must contain at least 10 meaningful characters."
-            )
-
-        reviewed_candidate_ids = Bibliography._normalize_reviewed_candidate_ids(
-            override.get("reviewedCandidateIds")
-        )
-        current_candidate_ids = {
-            candidate["id"] for candidate in duplicate_result["candidates"]
-        }
-        blocking_candidate_ids = Bibliography._blocking_duplicate_candidate_ids(
-            duplicate_result
-        )
-        if any(
-            candidate_id not in current_candidate_ids
-            for candidate_id in reviewed_candidate_ids
-        ):
-            raise DuplicateOverrideError(
-                "override.reviewedCandidateIds must match the current duplicate "
-                "candidates returned by the server-side rerun."
-            )
-        if not set(reviewed_candidate_ids).intersection(blocking_candidate_ids):
-            raise DuplicateOverrideError(
-                "override.reviewedCandidateIds must include at least one current "
-                "blocking duplicate candidate."
-            )
-
-    @staticmethod
-    def _blocking_duplicate_candidate_ids(
-        duplicate_result: Mapping[str, Any],
-    ) -> set[str]:
-        return {
-            candidate["id"]
-            for candidate in duplicate_result["candidates"]
-            if candidate["decision"] in BLOCKING_DUPLICATE_DECISIONS
-        }
-
-    @staticmethod
-    def _normalize_reviewed_candidate_ids(candidate_ids: Any) -> Sequence[str]:
-        if not isinstance(candidate_ids, Sequence) or isinstance(
-            candidate_ids, (str, bytes)
-        ):
-            raise DuplicateOverrideError(
-                "override.reviewedCandidateIds must be a non-empty array of strings."
-            )
-
-        normalized_candidate_ids = []
-        for candidate_id in candidate_ids:
-            if not isinstance(candidate_id, str) or not candidate_id.strip():
-                raise DuplicateOverrideError(
-                    "override.reviewedCandidateIds must be a non-empty array of strings."
-                )
-            normalized_candidate_ids.append(candidate_id.strip())
-
-        deduplicated_candidate_ids = list(dict.fromkeys(normalized_candidate_ids))
-        if not deduplicated_candidate_ids:
-            raise DuplicateOverrideError(
-                "override.reviewedCandidateIds must be a non-empty array of strings."
-            )
-        return deduplicated_candidate_ids
 
     def export_page(
         self, cursor: Optional[str] = None, limit: int = DEFAULT_EXPORT_LIMIT
