+++ ebl/bibliography/application/duplicate_audit_reports.py
@@ -1,0 +1,178 @@
+from __future__ import annotations
+
+import csv
+import json
+from pathlib import Path
+from typing import Any, Mapping, Sequence
+
+from ebl.bibliography.application.duplicate_audit_groups import CandidateGroup
+from ebl.bibliography.application.duplicate_audit_normalization import (
+    NormalizedBibliographyEntry,
+)
+from ebl.bibliography.application.duplicate_audit_scoring import PairScore
+from ebl.bibliography.application.duplicate_audit_summary import (
+    audit_statistics,
+    pair_to_dict,
+)
+
+
+def write_reports(
+    output_dir: Path,
+    entries: Sequence[NormalizedBibliographyEntry],
+    pairs: Sequence[PairScore],
+    groups: Sequence[CandidateGroup],
+) -> None:
+    output_dir.mkdir(parents=True, exist_ok=True)
+    stats = audit_statistics(entries, pairs)
+    (output_dir / "bibliography_duplicate_candidate_groups.json").write_text(
+        json.dumps([group.to_dict() for group in groups], indent=2, ensure_ascii=False),
+        encoding="utf-8",
+    )
+    write_pairs_csv(output_dir / "bibliography_duplicate_candidate_pairs.csv", pairs)
+    write_groups_csv(output_dir / "bibliography_duplicate_candidate_groups.csv", groups)
+    write_summary(output_dir / "bibliography_duplicate_audit_summary.md", stats, groups)
+
+
+def write_pairs_csv(path: Path, pairs: Sequence[PairScore]) -> None:
+    with path.open("w", encoding="utf-8", newline="") as stream:
+        writer = csv.DictWriter(stream, fieldnames=pair_csv_fields())
+        writer.writeheader()
+        for pair in pairs:
+            row = pair_to_dict(pair)
+            row["matchedSignals"] = json.dumps(row["matchedSignals"], sort_keys=True)
+            row["conflictingSignals"] = ",".join(row["conflictingSignals"])
+            writer.writerow(row)
+
+
+def pair_csv_fields() -> list[str]:
+    return [
+        "leftId",
+        "rightId",
+        "score",
+        "decision",
+        "matchedSignals",
+        "conflictingSignals",
+        "evidenceCompleteness",
+        "reason",
+        "recommendation",
+        "previouslyReviewedNotDuplicate",
+    ]
+
+
+def write_groups_csv(path: Path, groups: Sequence[CandidateGroup]) -> None:
+    with path.open("w", encoding="utf-8", newline="") as stream:
+        writer = csv.DictWriter(stream, fieldnames=group_csv_fields())
+        writer.writeheader()
+        for group in groups:
+            writer.writerow(group_csv_row(group.to_dict()))
+
+
+def group_csv_fields() -> list[str]:
+    return [
+        "candidateGroupId",
+        "groupDecision",
+        "highestScore",
+        "averageScore",
+        "memberIds",
+        "suggestedCanonicalId",
+        "canonicalReason",
+        "needsHumanReview",
+        "matchedSignalsSummary",
+        "conflictingSignalsSummary",
+    ]
+
+
+def group_csv_row(data: Mapping[str, Any]) -> dict[str, Any]:
+    return {
+        **{
+            key: data[key]
+            for key in (
+                "candidateGroupId",
+                "groupDecision",
+                "highestScore",
+                "averageScore",
+                "suggestedCanonicalId",
+                "canonicalReason",
+                "needsHumanReview",
+            )
+        },
+        "memberIds": ",".join(data["memberIds"]),
+        "matchedSignalsSummary": json.dumps(
+            data["matchedSignalsSummary"], sort_keys=True
+        ),
+        "conflictingSignalsSummary": ",".join(data["conflictingSignalsSummary"]),
+    }
+
+
+def write_summary(
+    path: Path, stats: Mapping[str, Any], groups: Sequence[CandidateGroup]
+) -> None:
+    path.write_text("\n".join(summary_lines(stats, groups)) + "\n", encoding="utf-8")
+
+
+def summary_lines(
+    stats: Mapping[str, Any], groups: Sequence[CandidateGroup]
+) -> list[str]:
+    lines = [
+        "# Bibliography Duplicate Audit Summary",
+        "",
+        "This report was generated by a read-only audit. It did not modify MongoDB records.",
+        "",
+        "## Calibration Metrics",
+        "",
+    ]
+    lines.extend(f"- `{key}`: {value}" for key, value in stats.items())
+    lines.extend(issn_section(groups))
+    lines.extend(group_section("Top Likely Duplicate Groups", likely_groups(groups)))
+    lines.extend(
+        group_section("Top Possible Duplicate Groups", possible_groups(groups))
+    )
+    lines.extend(recommendation_section())
+    return lines
+
+
+def issn_section(groups: Sequence[CandidateGroup]) -> list[str]:
+    return [
+        f"- `fuzzyDuplicateGroups`: {len(groups)}",
+        "",
+        "## ISSN Handling",
+        "",
+        "Exact ISSN groups are serial/series support groups only. ISSN alone does not create a hard likely-duplicate decision.",
+        "",
+    ]
+
+
+def likely_groups(groups: Sequence[CandidateGroup]) -> list[CandidateGroup]:
+    return [group for group in groups if group.group_decision == "likely_duplicate"]
+
+
+def possible_groups(groups: Sequence[CandidateGroup]) -> list[CandidateGroup]:
+    return [group for group in groups if group.group_decision == "possible_duplicate"]
+
+
+def group_section(title: str, groups: Sequence[CandidateGroup]) -> list[str]:
+    lines = [f"## {title}", ""]
+    lines.extend(group_markdown(group) for group in groups[:10])
+    lines.append("")
+    return lines
+
+
+def recommendation_section() -> list[str]:
+    return [
+        "## Recommended Thresholds",
+        "",
+        "Start review with `likely_duplicate >= 0.92` and `possible_duplicate >= 0.76`. Adjust after manual review of false positives, sparse records, and edition/variant cases.",
+        "",
+        "## Known Limitations",
+        "",
+        "- Note-markup usage is best-effort and may miss dynamically generated or differently serialized `@bib{}` references.",
+        "- The offline detector uses broader blocking than a future online API should use.",
+        "- No merge, deprecation, or tombstone workflow is implemented here.",
+    ]
+
+
+def group_markdown(group: CandidateGroup) -> str:
+    return (
+        f"- `{group.group_id}` {group.group_decision}, highest={group.highest_score}, "
+        f"members={', '.join(group.member_ids)}, canonical={group.suggested_canonical_id}"
+    )
