+++ ebl/tests/bibliography/test_partner_bibliography_override_validation_route.py
@@ -1,0 +1,194 @@
+import json
+
+import falcon
+import pytest
+
+from ebl.tests.bibliography.bibliography_route_test_helpers import (
+    NON_BLOCKING_DUPLICATE_CANDIDATE_ID,
+    STALE_DUPLICATE_CANDIDATE_ID,
+    assert_duplicate_override_bad_request,
+    blocking_duplicate_override_result,
+    client_with_scope,
+    duplicate_override_payload,
+    mixed_duplicate_override_result,
+    patch_duplicate_override_result,
+)
+from ebl.tests.factories.bibliography import BibliographyEntryFactory
+
+
+def test_partner_bibliography_duplicate_override_missing_override_does_not_mutate(
+    client, database, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        {"bibliographyEntry": duplicate_entry},
+    )
+
+
+def test_partner_bibliography_duplicate_override_whitespace_reason_does_not_mutate(
+    client, database, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        duplicate_override_payload(duplicate_entry, [saved_entry["id"]], "   "),
+        description_substring="override.reason",
+    )
+
+
+def test_partner_bibliography_duplicate_override_short_reason_does_not_mutate(
+    client, database, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        duplicate_override_payload(duplicate_entry, [saved_entry["id"]], "too short"),
+        description_substring="10 meaningful characters",
+    )
+
+
+@pytest.mark.parametrize(
+    "payload",
+    [
+        lambda entry, candidate_id: {
+            "bibliographyEntry": entry,
+            "override": {"reason": "Reviewed returned duplicate candidates carefully."},
+        },
+        lambda entry, candidate_id: duplicate_override_payload(entry, []),
+    ],
+)
+def test_partner_bibliography_duplicate_override_requires_reviewed_candidate_ids(
+    payload, client, database, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        payload(duplicate_entry, saved_entry["id"]),
+    )
+
+
+def test_partner_bibliography_duplicate_override_rejects_stale_candidate_ids(
+    client, database, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        duplicate_override_payload(duplicate_entry, ["Q39999999"]),
+        description_substring="must match the current duplicate candidates",
+    )
+
+
+def test_partner_bibliography_duplicate_override_rejects_only_non_blocking_candidate_id(
+    monkeypatch, client, database
+):
+    patch_duplicate_override_result(monkeypatch, mixed_duplicate_override_result())
+    bibliography_entry = BibliographyEntryFactory.build(id="Q30000001")
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        duplicate_override_payload(
+            bibliography_entry, [NON_BLOCKING_DUPLICATE_CANDIDATE_ID]
+        ),
+        description_substring="current blocking duplicate candidate",
+    )
+
+
+def test_partner_bibliography_duplicate_override_rejects_stale_candidate_after_rerun(
+    monkeypatch, client, database
+):
+    patch_duplicate_override_result(monkeypatch, blocking_duplicate_override_result())
+    bibliography_entry = BibliographyEntryFactory.build(id="Q30000001")
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        duplicate_override_payload(bibliography_entry, [STALE_DUPLICATE_CANDIDATE_ID]),
+        description_substring="must match the current duplicate candidates",
+    )
+
+
+def test_partner_bibliography_duplicate_override_unique_entry_uses_normal_create(
+    client, database
+):
+    bibliography_entry = BibliographyEntryFactory.build(id="Q30000001")
+    before_count = database["bibliography"].count_documents({})
+
+    result = client.simulate_post(
+        "/api/v1/bibliography/duplicate-override",
+        body=json.dumps(duplicate_override_payload(bibliography_entry, ["Q39999999"])),
+    )
+
+    assert result.status == falcon.HTTP_BAD_REQUEST
+    assert "Use POST /api/v1/bibliography instead." in result.json["description"]
+    assert database["bibliography"].count_documents({}) == before_count
+
+
+def test_partner_bibliography_duplicate_override_invalid_csl_does_not_mutate(
+    client, database
+):
+    assert_duplicate_override_bad_request(
+        client,
+        database,
+        {
+            "bibliographyEntry": {"id": "Q30000001", "title": "Missing type"},
+            "override": {
+                "reason": (
+                    "Reviewed returned duplicate candidates and confirmed this is "
+                    "a distinct bibliography record."
+                ),
+                "reviewedCandidateIds": ["Q30000000"],
+            },
+        },
+    )
+
+
+def test_partner_bibliography_duplicate_override_requires_write_scope(
+    guest_client, saved_entry
+):
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+
+    result = guest_client.simulate_post(
+        "/api/v1/bibliography/duplicate-override",
+        body=json.dumps(
+            duplicate_override_payload(duplicate_entry, [saved_entry["id"]])
+        ),
+    )
+
+    assert result.status == falcon.HTTP_FORBIDDEN
+
+
+def test_partner_bibliography_duplicate_override_rejects_duplicate_check_only_scope(
+    context, saved_entry
+):
+    client = client_with_scope(context, "check:bibliography_duplicates")
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+
+    result = client.simulate_post(
+        "/api/v1/bibliography/duplicate-override",
+        body=json.dumps(
+            duplicate_override_payload(duplicate_entry, [saved_entry["id"]])
+        ),
+    )
+
+    assert result.status == falcon.HTTP_FORBIDDEN
+
+
+def test_partner_bibliography_duplicate_override_rejects_export_only_scope(
+    context, saved_entry
+):
+    client = client_with_scope(context, "export:bibliography")
+    duplicate_entry = {**saved_entry, "id": "Q30000001"}
+
+    result = client.simulate_post(
+        "/api/v1/bibliography/duplicate-override",
+        body=json.dumps(
+            duplicate_override_payload(duplicate_entry, [saved_entry["id"]])
+        ),
+    )
+
+    assert result.status == falcon.HTTP_FORBIDDEN
